refector : 에러 및 로딩페이지 수정
