--- client/src/components/mypage/Loading.tsx
@@ -1,7 +1,0 @@
-import loadingImg from '../../images/loading.gif';
-
-const Loading = () => {
-  return <img src={loadingImg} alt="로딩중" style={{ width: '52px' }} />;
-};
-
-export default Loading;
