--- client/src/components/mypage/NoSearchResult.tsx
@@ -1,5 +1,0 @@
-const NoSearchResult = () => {
-  return <div>문제집이 없습니다.</div>;
-};
-
-export default NoSearchResult;
